backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.0.0
PyJWT==2.8.0
redis==4.5.4
mysql-connector-python==8.1.0
Flask-Session==0.5.0